refactor: add guard clause for existing id to deleteArticle in the service implication

+++ server/src/main/java/org/vgr/server/service/impl/ArticleServiceImpl.java
@@ -33,8 +33,8 @@ public class ArticleServiceImpl implements ArticleService {
 
     @Override
     public void deleteArticle(Long id) {
-        articleRepository.deleteById(id);
-
+        if(!articleRepository.existsById(id)) throw new ArticleNotFoundException(id);
+         articleRepository.deleteById(id);
     }
 
     @Override
